+++ include/MemeAudio/AudioSource.hpp
@@ -17,10 +17,10 @@ namespace ml
 		~AudioSource();
 
 		AudioSource & clean();
-		AudioSource & create(size_t count);
+		AudioSource & create(uint32_t count);
 
 	private:
-		size_t m_count;
+		uint32_t m_count;
 	};
 }
